Added Login and Logout functionalities

+++ app/Controller/PostsController.php
@@ -3,8 +3,9 @@ class PostsController extends AppController {
 	public $helpers = array('Html', 'Form', 'Flash');
 	public $components = array('Flash');
 
-	public function index(){
-		$this->set('posts', $this->Post->find('all'));
+	public function index() {
+		//$this->set('posts', $this->Post->find('all'));
+		$this->set('posts', $this->Post->find('all', ['Post.user_id' => 'User.id', 'recursive' => 1]));
 	}
 
 	//File Name: view.ctp
@@ -20,20 +21,24 @@ class PostsController extends AppController {
 		}
 		$this->set('post', $post);
 	}
+
 	public function add() {
-		//This is to check a form method
-		if ($this->request->is('post')) {
-			//to save new information
-			//$this->Post->create();
-			$this->request->data['Post']['user_id'] = $this->Auth->user('id');
-			//check the validation error and check if it needs to be stopped
-			if ($this->Post->save($this->request->data)) {
-				//$this->request->data contains data from the post method
-				//Flash method sets the success message into the session variable, allowing it to be displayed on the redirected page.
-				$this->Flash->success(__('Your post has been saved'));
-				return $this->redirect(array('action' => 'index'));
+		if (!empty($this->Session->read('Auth.User.User.id'))) {
+			//This is to check a form method
+			if ($this->request->is('post')) {
+				//to save new information
+				$this->Post->create();
+				//manipulate the post data so that it can be inserted into the database
+				//$this->Post->save = INSERT INTO POST ...
+				$this->request->data['Post']['user_id'] = $this->Session->read('Auth.User.User.id');
+				if ($this->Post->save($this->request->data)) {
+					//$this->request->data contains data from the post method
+					//Flash method sets the success message into the session variable, allowing it to be diplayed on the redirected page.
+					$this->Flash->success(__('Your post has been saved'));
+					return $this->redirect(array('action' => 'index'));
+				}
+				$this->Flash->error(__('Unable to add your post.'));
 			}
-			$this->Flash->error(__('Unable to add your post.'));
 		}
 	}
 
@@ -43,17 +48,23 @@ class PostsController extends AppController {
 		}
 
 		$post = $this->Post->findById($id);
+		$user_id = $post['Post']['user_id'];
 		if (!$post) {
 			throw new NotFoundException(__('Invalid Post'));
 		}
 
-		if ($this->request->is(array('post', 'put'))) {
-			$this->Post->id = $id;
-			if ($this->Post->save($this->request->data)) {
-				$this->Flash->success(__('Your post has been updated'));
-				return $this->redirect(array('action' => 'index'));
+		if ($user_id == $this->Session->read('Auth.User.User.id')) {
+			if ($this->request->is(array('post', 'put'))) {
+				$this->Post->id = $id;
+				if ($this->Post->save($this->request->data)) {
+					$this->Flash->success(__('Your post has been updated'));
+					return $this->redirect(array('action' => 'index'));
+				}
+				$this->Flash->error(__('Unable to update your post'));
 			}
-			$this->Flash->error(__('Unable to update your post'));
+		} else {
+			$this->Flash->error(__('Unable to access to the post'));
+			$this->redirect(array('action' => 'index'));
 		}
 
 		if (!$this->request->data) {
@@ -66,13 +77,41 @@ class PostsController extends AppController {
 			throw new MethodNotAllowedException();
 		}
 
-		if ($this->Post->delete($id)) {
-			$this->Flash->success(__('The post with id : %s has been deleted.' , h($id)));
+		$post = $this->Post->findById($id);
+
+		if (!$post) {
+			throw new NotFoundException(__('Invalid Post'));
+		}
+
+		$user_id = $post['Post']['user_id'];
+
+		if ($user_id == $this->Session->read('Auth.User.User.id')) {
+
+			if ($this->Post->delete($id)) {
+				$this->Flash->success(__('The post with id : %s has been deleted.' , h($id)));
+			} else {
+				$this->Flash->error(__('The post with id : %s could not be deleted' . h($id)));
+			}
 		} else {
-			$this->Flash->error(__('The post with id : %s could not be deleted' . h($id)));
+			$this->Flash->error(__('Unable to access to the post'));
 		}
 
 		return $this->redirect(array('action' => 'index'));
+	}
+
+	public function isAuthorized($user) {
+		if ($this->action == 'add') {
+			return true;
+		}
+
+		if (in_array($this->action, array('edit', 'delete'))) {
+			$postId = (int) $this->request->params['pass'][0];
+			if ($this->Post->isOwnedBy($postId, $user['id'])) {
+				return true;
+			} else {
+				throw new MethodNotAllowedException('他人の投稿は編集できません');
+			}
+		}
 	}
 }
